# config/documentation/swagger-admin.json
{
	"swagger": "2.0",
	"info": {
		"title": "HonorVet - Admin API",
		"description": "HonorVet - Admin API Documentation",
		"version": "1.0"
	},
	"produces": ["application/json"],
	"basePath": "/",
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "authorization",
			"in": "header"
		}
	},
	"paths": {
...
		"/admin/v1/admin/": {
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-put",
				"tags": ["Sub Admin"],
				"description": "Update Admin Details",
				"security": [
					{
						"Bearer": []
					}
				],
				"content": "multipart/form-data",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"properties": {
								"id": {
									"description": "id GUID",
									"required": false,
									"type": "string",
									"value": ""
								},
								"email": {
									"description": "email",
									"required": false,
									"type": "string",
									"value": ""
								},
								"firstName": {
									"description": "First Name",
									"required": false,
									"type": "string",
									"value": ""
								},
								"lastName": {
									"description": "Last Name",
									"type": "string",
									"value": ""
								},
								"adminType": {
									"enum": ["SUPER_ADMIN", "SUB_ADMIN"],
									"description": "",
									"type": "string",
									"required": false
								},
								"accessPermissions": {
									"type": "array",
									"required": false,
									"items": {
										"$ref": "#/definitions/accessPermissions"
									}
								}
							}
						}
					}
				],
				"responses": {}
			},
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-post",
				"security": [
					{
						"Bearer": []
					}
				],
				"tags": ["Sub Admin"],
				"description": "Add SuperAdmins or SubAdmins",
				"consumes": "application/json",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"properties": {
								"email": {
									"description": "email is required",
									"required": true,
									"type": "string",
									"value": ""
								},
								"firstName": {
									"description": "First Name",
									"required": true,
									"type": "string",
									"value": ""
								},
								"lastName": {
									"description": "Last Name",
									"type": "string",
									"value": ""
								},
								"adminType": {
									"enum": ["SUPER_ADMIN", "SUB_ADMIN"],
									"description": "",
									"type": "string",
									"required": true
								},
								"accessPermissions": {
									"type": "array",
									"required": true,
									"items": {
										"$ref": "#/definitions/accessPermissions"
									}
								}
							}
						}
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-delete",
				"tags": ["Sub Admin"],
				"description": "Delete Admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/admin/block": {
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-block-put",
				"tags": ["Sub Admin"],
				"description": "Block Admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"required": true
					},
					{
						"name": "isBlocked",
						"in": "formData",
						"description": "",
						"enum": [0, 1],
						"type": "integer",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/skills/": {
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminskills-create-POST",
				"tags": ["Skills"],
				"description": "Add skills by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": true
					}
				],
				"responses": {}
			},
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminskills-list-get",
				"tags": ["Skills"],
				"description": "Get skills by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			},
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminSkills-update-put",
				"tags": ["Skills"],
				"description": "Update skills by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": false
					},
					{
						"name": "isBlocked",
						"in": "formData",
						"description": "",
						"type": "number",
						"required": false
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminskills-delete-delete",
				"tags": ["Skills"],
				"description": "Delete skills by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/skills/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminsskills-detail-{id}-get",
				"tags": ["Skills"],
				"description": "Get single skills by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/schedules/": {
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminSchedules-create-POST",
				"tags": ["Schedules"],
				"description": "Add Schedules by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": true
					}
				],
				"responses": {}
			},
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminSchedules-list-get",
				"tags": ["Schedules"],
				"description": "Get Schedules by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			},
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminSchedules-update-put",
				"tags": ["Schedules"],
				"description": "Update Schedules by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": false
					},
					{
						"name": "isBlocked",
						"in": "formData",
						"description": "",
						"type": "number",
						"required": false
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminSchedules-delete-delete",
				"tags": ["Schedules"],
				"description": "Delete Schedules by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/schedules/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminSchedules-detail-{id}-get",
				"tags": ["Schedules"],
				"description": "Get single Schedules by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/users/listing": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-users-listing-get",
				"tags": ["Users Management"],
				"description": "Get User listing by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "startDate",
						"in": "query",
						"description": "YYYY-MM-DD",
						"required": false,
						"type": "",
						"value": ""
					},
					{
						"name": "endDate",
						"in": "query",
						"description": "YYYY-MM-DD",
						"required": false,
						"type": "date",
						"value": ""
					},
					{
						"name": "isBlocked",
						"enum": [0, 1],
						"description": "0 => unBlocked Users , 1=> Blocked Users",
						"in": "query",
						"type": "number",
						"required": false
					},
					{
						"name": "skillId",
						"in": "formData",
						"description": "['10','11','12']",
						"type": "string",
						"value": "",
						"required": false
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/users/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-user-detail-{id}-get",
				"tags": ["Users Management"],
				"description": "Get User details by admin",

				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/users/delete": {
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-user-{id}-delete-delete",
				"tags": ["Users Management"],
				"description": "Delete User by admin",

				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/users/block": {
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-user-{id}-block",
				"tags": ["Users Management"],
				"description": "block User by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "isBlocked",
						"enum": [0, 1],
						"description": "Blocked admin users",
						"in": "formData",
						"type": "integer",
						"required": false
					}
				],
				"responses": {}
			}
		},
...
	},
	"definitions": {
		"accessPermissions": {
			"type": "object",
			"properties": {
				"module": {
					"type": "string"
				},
				"permission": {
					"enum": [1, 0],
					"description": "0 or 1 to set permission",
					"type": "integer"
				}
			},
			"example": {
				"module": "dashboard",
				"permission": 1
			}
		}
	}
}